requirements.txt
asyncpg==0.29.0
icecream==2.1.3
nextcord==2.6.0
python-dotenv==1.0.1
python_bcrypt==0.3.2
